Differentiate local imports with tilda
