+++ archetypes/Stake/FarmsTable/index.tsx
@@ -8,7 +8,7 @@ import { TWModal } from '@components/General/TWModal';
 import Close from '/public/img/general/close.svg';
 import { Logo, LogoTicker, tokenSymbolToLogoTicker } from '@components/General/Logo';
 import Loading from '@components/General/Loading';
-import { BalancerPoolAsset } from '@libs/types/Staking';
+import { BalancerPoolAsset } from '~/types/staking';
 import { calcAPY, calcBptTokenPrice } from '@tracer-protocol/pools-js';
 import { APYTip, RewardsEndedTip } from '@components/Tooltips';
 import { TokenToFarmAddressMap } from '@constants/farms';
