+++ archetypes/Stake/StakeGeneric/index.tsx
@@ -18,7 +18,7 @@ import { FilterFilled, SearchOutlined } from '@ant-design/icons';
 import { useWeb3 } from '@context/Web3Context/Web3Context';
 import FarmNav from '@components/Nav/FarmNav';
 import StakeModal from '../StakeModal';
-import { Farm } from '@libs/types/Staking';
+import { Farm } from '~/types/staking';
 import { Logo, LogoTicker } from '@components/General/Logo';
 import { useStore } from '@store/main';
 import { TransactionType } from '@store/TransactionSlice/types';
