+++ constants/bridge.ts
@@ -1,7 +1,7 @@
 import { tokenMap } from '@tracer-protocol/pools-js/data';
 import { ARBITRUM_RINKEBY, ARBITRUM, MAINNET, RINKEBY } from '@constants/networks';
 import { LogoTicker } from '@components/General/Logo';
-import { BridgeableAssets, BridgeableAssetWarnings, DestinationNetwork } from '@/types/bridge';
+import { BridgeableAssets, BridgeableAssetWarnings, DestinationNetwork } from '~/types/bridge';
 
 export const destinationNetworkLookup: Record<string, DestinationNetwork> = {
     [RINKEBY]: ARBITRUM_RINKEBY,
