+++ context/SwapContext.tsx
@@ -1,7 +1,8 @@
 import React, { useContext, useReducer, useMemo, useEffect } from 'react';
 import BigNumber from 'bignumber.js';
 import { CommitActionEnum, BalanceTypeEnum, SideEnum } from '@tracer-protocol/pools-js';
-import { Children, PoolType } from '@libs/types/General';
+import { Children } from '~/types/general';
+import { PoolType } from '~/types/pools';
 import { useRouter } from 'next/router';
 import { usePools } from './PoolContext';
